pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dataclasses-excel"
dynamic = ["version"]
description = 'Serialize dataclasses from Excel spreadsheets'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Rob Thatcher", email = "rob@tin.works" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  "openpyxl",
  "pytest-datadir",
  "rich"
]

[project.urls]
Documentation = "https://github.com/thatcr/dataclasses-excel#readme"
Issues = "https://github.com/thatcr/dataclasses-excel/issues"
Source = "https://github.com/thatcr/dataclasses-excel"

[project.scripts]
dataclasses-excel = "dataclasses_excel.cli:dataclasses_excel"

[tool.hatch.version]
path = "src/dataclasses_excel/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/dataclasses_excel tests}"

[tool.coverage.run]
source_pkgs = ["dataclasses_excel", "tests"]
branch = true
parallel = true
omit = [
  "src/dataclasses_excel/__about__.py",
]

[tool.coverage.paths]
dataclasses_excel = ["src/dataclasses_excel", "*/dataclasses-excel/src/dataclasses_excel"]
tests = ["tests", "*/dataclasses-excel/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.ruff.lint]
